Fix spock_apply_heap.c includes for PG15 and usage (lack thereof).
The header file utils/int8.h was removed as part of the commit
postgres/postgres@cfc7191df.

However, the fix isn't to conditionally include it.  This header was
included but the code does not use any declarations (scanint8) from it.

So we simply remove it as it is unused in this source file.