Fix for PG15 cross-partition UPDATE foreign key correctness changes.
See postgres/postgres@ba9a7e39.

We are following the lead of the logical replication code in PostgreSQL
core which presently passes in NULLs and false for the parameters
pertaining to cross-partition UPDATE trigger queuing.

By getting through the compilation errors and producing a dynamic
library, we will gain the ability to start testing our changes.

We can test cross-partition UPDATEs following that and prioritize
accordingly thereafter.