Adding patch that needs to be applied to core PG
This patch implements the attribute option LOG_OLD_VALUE.
When one or more attributes of a relation are marked this
way, heap_update() is forced to create an old-key tuple
and include these attributes in it. This makes the old
values available in the output plugin and in the case of
Spock, automatically available in spock_heap_apply_update().

Configuring a column is done by

    ALTER TABLE <tablename>
        ALTER COLUMN <columnname> SET(LOG_OLD_VALUE=true);