First patch for delta apply of balance/counter type columns
Use the attribute option LOG_OLD_VALUE as an indicator to
apply remote changes as a delta. This is currently limited
to the data types int4, int8 and numeric. Support for more
datatypes will be added before we release this feature.

WARNING: At this point the patch breaks last_update_wins
logic by making it possible that the commit_timestamp of
a row goes backwards. We are actively working on a solution
for this problem.