First part of conflict tracking
The delta-resolution introduced in 2209106e301 requires
tracking of the correct last commit timestamp for rows,
where we force an update to delta columns. This patch
adds a shared hash table that implements this.

Under heavy load for long enough the hash table will
eventually run out of entries to track. The next step
is to implement purging functionality that cleans up
the hash table from known to be no longer needed entries.