Avoid pruning when we don't have every last_ts
We cannot safely prune conflict tracking data unless
we have a last commit timestamp for every subscription.
Unfortunately we only learn that information (at this
point) when receiving data for a subscription in the
corresponding apply worker process. This means that we
cannot prune when a node is down or unreachable.