Add more data types to DCA
This adds INT2, FLOAT4, FLOAT8, MONEY and BOOL to the
data types, that delta conflict avoidance (DCA) can
handle.

The logic for BOOL is basically OR. Meaning that we
always end up with a TRUE value whenevern any node
transitions the value from FALSE to TRUE. We can never
go back to FALSE without disabling DCA for the column.