Elevate access level in spock_object_access()
This function is a hook callback for ProcessUtility.
It does a dependency check for which it needs access
to the spock schema and tables but it is called under
the current session user, who might not have those
privileges.

We elevate our access privileges to the BOOTSTRAP
user ID, which is a superuser. This is done in a way
that automatically resets the changes on transaction
abort.