lag_tracker renaming for consistency
Renaming the lagtracker to lag_tracker for consistency
with other functions and views. Same for the column
names.

Removed unused columns from the lag_tracker_info()
function. Based the lag time in the view on timeofday()
because the lag can otherwise display negative if there
is a clock difference between the servers.