Remove conflict_tracking_hash
The conflict tracking shared memory hash table
was originally the result of delta-apply causing
anomalies with last-update-wins resolution. The
forced last-update-wins tuple-updated would make
core-PG report the new timestamp instead of the
correct previous one, so we had to implement an
override mechanism.

The resulting shared memory hash table turned out
to be prone to out-of-memory conditions as well
as not being postmaster crash safe, so we then
added a permanent table as backing store and
pruned the shmem hash table on every transaction
begin event.

Now we have another case where the shmem hash
table can overflow, so it is time to ditch it.