Wait for synchronous replication when decoding COMMIT
When our node is configured for Ultra-High-Availability
we have a synchronous standby configuration. We must wait
for the quorum to be satisfied before sending a COMMIT
message from the walsender to the apply_worker. Otherwise
a failover to the synchronous standby can lose a local
transaction that was already sent to another Spock node.