Small fix to wait_for_sync_rep_in_commit
The LSN to wait for via SyncRepWaitForLSN() is not the
commit_lsn, but the end_lsn in the ReorderBufferTXN.
That LSN is the position AFTER the commit record.
Using the commit_lsn caused the walsender to issue
a keepalive ping, to which the recovery worker would
reply that it had applied everything up to the commit_lsn.
But of course not the commit record itself yet.