Ensure that the LSN for SyncRepWaitForLSN() is unique
A current problem in SyncRepWaitForLSN() causes an Assert()
failure. The existing core code assumes that there will only
ever be one proc waiting for a specific LSN. We will use our
slot number in the walsender array and subtract that from
the end_lsn. This will work as long as we never have more
walsenders than a commit message is long (minimum currently
is 48 bytes).