Revert 2b3424c83c526e9fc9dd36d9e73ec4a6ce15dd5f
This reverts the attempt to create roles and security
definer functions to support multi-tenant operation
using multiple databases. The attempt was flawed and
created many problems. The Spock administrator and
replication user needs to be a real PostgreSQL
superuser.

This removes the two roles spock_admin and spock_monitor.
We might want to recreate the spock_monitor role with
specific GRANTs later to support the idea of a readonly
user for systems like Prometheus.