Fixing the "relcache reference leak" WARNING
This is a problem that surfaced due to an inquiry from one
of our Design Partners. When a user trigger is configured
as ENABLED ALWAYS, so that it fires during replication (in
this case to create a NOTIFY), we experienced a WARNING
about a "relcache reference leak".

It turns out that when there are triggers to fire because
of the above configuration, our code in spock_apply_heap.c
calls EvalPlanQualInit() and EvalPlanQualEnd(). I am not
yet sure those calls are necessary since our code never
calls EvalPlanQual(). But the fact is that Init() is
creating another relcache reference while End() is leaving
that reference behind (according to the comments in PG) on
the result relation. Our code is therefore responsible for
removing that reference by closing the result relation.