Part 1 of the conflict tracking redesign
Conflict tracking is an issue that arose from implementing
delta-apply. When an incoming remote tuple loses the
last-update-wins logic, the delta columns still need to be
updated. That "forced" update would associate the new tuple
with the wrong last-update timestamp, so another update
being replicated from a third node, that happens to commit
before the last local update but after the other remote one
would be applied in error.

This commit changes the initial attempt to keep track of
the relevant timestamps by storing them in a special user
catalog table to keeping them in "hidden/invisible" columns
of the row itself.

There is an essential part missing so far. The relevant
"invisible" columns need to be automatically added to a
table when it is added to a replication set and doesn't
have them yet. This will be done in a separate commit.