Implement a Lamport Clock for the commit timestamp
The majority of this commit is a patch to Core PG16 and
PG15 implementing a Lamport Clock that guarantees that
the commit timestamp in WAL records is monotonically
increasing in LSN order and that a remote transaction,
executed by the Spock apply worker, will bump the Lamport
clock into the future if the local system clock is lagging
behind the remote Lamport clock.

The Lamport Clock itself is a TimestampTz in the XlogCtl
structure held in shared memory. A new hook is used to
call our custom adjustment function during the time, the
WAL-insert spinlock is held. This is the spinlock during
which the LSN of the commit record is allocated, which
guarantees that there can be no more race conditions that
would cause the commit timestamp to become stale or run
backwards (as is the case in core PostgreSQL).