Limit node.id to 16 bits and extract tiebreaker
The plan in this branch is to replace the RepOriginId
with an ID that represents the actual origin of the
node, where the transaction initially took place,
instead of an identifier that only leads to the current
provider of a subscription that may be reused after
dropping and recreating subscriptions.

In this first commit the node.id is simply truncated
from the original 32-bit CRC of the node.name. In
addition we extract an optionally set "tiebreaker" value
from the node.info jsonb column that will later be used
in last-update-wins conflict resolution.