Fix transaction origin
For transactions that originated locally (txn->origin_id == 0)
we transmit our own node.id in the Origin message. For transactions
that we are forwarding we use the origin_id provided by
logical decoding, which is what we got from the provider. This
means that we can now identify where a transaction really
originated, instead of who the provider was (which could be a
forwarder).

In the apply-worker we set the global replorigin_session_origin,
which contained the roident identifying the slot as set by logical
decoding, to the received node.id. This causes the commit timestamp
tracking function TransactionIdGetCommitTsData() to return the
node.id of the true origin of the transaction, forwarded or not.

This required a protocol change. Specifically the format of the
ORIGIN wire message changed from transmitting the node.name of the
transaction origin to the node.id instead. We currently do not
plan on supporting a rolling upgrade from v3.x to v4.x because
of several other incompatibilities between the major versions.
If we support an upgrade in place at all it most likely will need
a brief maintenance window.