Changes for PG17 (#38)
* Adjust backend patches

Add new backend patches and adjust all of them
for fuzz/offset so they apply cleanly to PG 17devel.

* PG17 compat MemoryContextReset

Commit 6a72c42fd removed macro MemoryContextResetAndDeleteChildren().
The same is always done in MemoryContextReset().

* PG17 compat InitResultRelInfo() is obsolete

Macro InitResultRelInfo() is obsolete and was
removed from compat16 after compat17 was created
from it.

* PG17 compat obsolete tuplestore_donestoring()

Commit 75680c3d8 removed the obsolete macro
tuplestore_donestoring().

* PG17 compat flush option for LogLogicalMessage()

Commit 173b56f1e added a flush option to function
LogLogicalMessage().

* PG17 compat obsolete ExecInitExtraTupleSlot()

Macro ExecInitExtraTupleSlot() is obsolete and was
removed from compat16 after compat17 was created
from it.

* PG17 compat restore type ObjectClass

Commit 89e5ef7e2 removed the enum type ObjectClass.
Instead of adding multiple #ifdef PG_VERSION_NUM cases
we rather recreate that type for now.

* PG17 compat obsolete EvalPlanQualInit()

Macro EvalPlanQualInit() is obsolete and was
removed from compat16 after compat17 was created
from it.

* PG17 compat remove obsolete SPI_push/pop()

Commit 75680c3d8 removed the obsolete NOOP macros
SPI_push() and SPI_pop().

* Revert "PG17 compat restore type ObjectClass"

This reverts commit fac27033bb708c5fb198d615db5c7574236fc954.

* PG17 compat removes type ObjectClass

Commit 89e5ef7e2 removed the enum type ObjectClass.
Removed the usage of ObjectClass and the array containing
Oid to ObjectClass map as it is no longer required. Verified
that this change works for PG14 as well.

* PG17 updated how HeapTuple is accessed for ReorderBufferChange

Commit 08e6344fd removed the ReorderBufferTupleBuf struct.
Added the function "ReorderBufferChangeHeapTuple" that takes
in the ReorderBufferChange ptr along with the type of tuple
desired, and returns associated HeapTuple.

NOTE: HeapTuple in the ReorderBufferChange struct must only
be accessed via this function interface.

PG16 compat updated.
PG15 compat updated.
PG14 compat updated.

* PG17 renaming standby_slot_names to pg_standby_slot_names

Commit bf279ddd1 introduced a GUC with the name standby_slot_names.
This was conflicting with our variable. Hence, it was renamed to
pg_standby_slot_names. Ideally this should be renamed to
spock_failover_slot_names_list or spock_failover_slot_names, but
those names are already taken. So, prefixed pg_ to the name to avoid
the conflict.

* PG17 updating ReplicationSlotCreate call to match new API

Commit c393308b69d introduced argument "bool failover".
commit ddd5f4f54a0 introduced argument "bool synced".

Both of these arguments are passed as false as it seems that
spock is doing failover slot handling itself. That may need
a more holistic review and change.

* PG17 core patch fixes Makefile.global.in to not set INITDB_TEMPLATE
in case when NO_TEMP_INSTALL is set.

Commit 252dcb32397 introduced initdb caching to speed up regression
testing. There seems no way to bypass this caching or unset the
INITDB_TEMPLATE variable when PGXS is used. This patch to the core
resolves this whilst perserving initdb caching.

We should share this patch with the community if it seems reasonable.

* Fix patch format to be compatible with `git apply'

* Adjust patch for line fuzz

* Disable FAILOVER keyword for now

For some reason PG17 activated a section of Spock code
that apparently issued some special keyword that is expected
by pglogical as implemented by 2nd Quadrant. We need to
investigate what exactly the purpose of this is, but it
isn't part of PG17. Only the existence of a certain function
triggered this.

---------

Co-authored-by: Jan Wieck <jan@wi3ck.info>