Refactor jsonb generation for exception handling
With PG17 a lot of the jsonb support functions have been
changed from static to extern. At the same time their
call signatures changed as well as the call signatures of
previously extern functions. This has caused our approach
of copying the needed jsonb functions into Spock to break.

This patch changes the way we build the jsonb data for the
exception_log table. The individual column data and names
are passed via SPI through to_json() for proper escaping.
This is a documented and SQL level exposed function, so we
can expect long-term stability for it. The remaining json
structure is very simple and is assembled via StringInfo
buffers. The resulting json string is then converted into
jsonb by calling the jsonb_in() function.

TODO: Separate the tuple data from the exception context
column so that the latter is reserved for things like DDL
errors.