Step 1 of moving delta-apply commit-ts into core
This creates a new function in core access/commit_ts that
can be used inside a subtransaction to record a separate
CommitTsData entry. The Spock apply worker is using this
by performing a forced delta-apply in a subtransaction and
recording the individual CommitTsData for it. Later on a
subsequent conflict resulution on the row in question will
retrieve this information intead of that from the toplevel
transaction by calling TransactionIdGetCommitTsData(). This
works because each row created by a subtransaction has its
own xmin and thus can have its own CommitTsData.

This commit also removes using hidden columns.

TODO: WAL logging the CommitTsData override information and
rebuilding the in-memory information on redo.