Protecting allocation of exception log slot with an LWLock.
At the moment, this is the only place where an exception log slot
is assigned. Therefore, we must guard it with an SpockCtx->lock.
Therefore, initial search to find an existing matching slot does
not need to be protected.

This is only a partial fix as we still need to perform cleanup
of these slots when no longer needed so that those may be reassigned.