Add upgrade script for Spock extension from 3.3 to 4.0
This upgrade script updates the node_id in spock.node, which is
referenced by multiple tables that do not have the ON UPDATE CASCADE
option enabled. To simplify this update, the constraints are
temporarily dropped and then re-added with the ON UPDATE CASCADE
option.

Additionally, several deprecated objects have been removed and a few
new ones have been added.