Stamp v4.0.2
Change version to v4.0.2 including schema upgrade script. This
includes a schema change for the exception_log* tables. They
need the retry_errored_at column in the PK because after a
SUB_DISABLE happened, the following sub_enable() could lead to
another set of exceptions (if the fix was incomplete) and then
the exception handling would create duplicate keys on retry.