Implementation of Commit Ordering Without Deadlock Detection
Commit ordering is assured by ensuring that the previous commit from the same
origin has been applied on the subscriber before the current transaction
applies its changes. This requires changes in the protocol and addition of
shared memory segment.

The output plugin now sends a new message 'L' that contains timestamps of the
last commited transaction prior to the current one. This message is sent
immediately after sending the origin information so that the subscriber can
associate the previous commit timestamp with an origin. This enables us to
sequence our commits when applying on the subscriber.

An apply worker recieves the 'L' message and proceeds on with recording the
previous commit timestamp locally. It then proceeds on with the transaction
processing as normal. When it receives the commit message, it sleeps if the
previous commit hasn't been applied. When a transaction commits, it broadcasts
to all waiters to wake up and check if their required commit has been
processed. If so, they proceed on with the commit processing otherwise they
go back to sleep.

A user catalog table "spock.progress" has been added that maintains latest
commited timestamp for a target and repl origin combination.

A shared memory segment is added for recording status of a apply groups
(apply workers with the same dbid and origin).

*** Test Case ***
To verify this behaviour, we had to patch spock to record all the commit
timestamps in a transaction table. The patch is available under the tests
folder.

There are C code changes as well as SQL changes where "progress_log" table
is added to record all commit timestamp changes in the progress table,
as well recording the sleep status of an apply worker.