Added a progress_log table to verifying commit ordering
*** NOT FOR PRODUCTION - FOR TESTING ONLY ***

progress_log is a transaction table that contains one row each for an apply
worker when:
	1.  it sleeps
	2.  when it is updated along with a sleep count

It also contains the first entry for a given origin so that we can track
the commit order from there onwards.

The table allows us to see the apply worker behaviour and validate that
they are proceeding in their respective origin/ts commit order.