Add exception_behaviour=sub_disable support
When the exception_behaviour is set to sub_disable, an
exception causes the apply-worker to disable the subscription.
The exception_log entries will include a message that tells
the DBA an LSN that can be used with the function
spock.sub_alter_skiplsn() to skip the transaction.
Alternatively the DBA can use spock.repair_mode() to perform
any manual changes that will enable the transaction to succeed.
Finally the DBA will use spock.sub_enable() to resume normal
operations.