In case of an update or a delete, if we don't find the local tuple, we should wait for the predecessor insert commit.
If that insert is never coming through, i.e. the tuple was locally
deleted, then retry will exceed the loop limit and we end up throwing
an ERROR that's not currently captured by the exception log.

An ERROR will cause the apply worker to exit without waking up
the waiters or updating it's timestamp. So, the sleeper will never
wake up.

TODO:
We need to handle the ERROR scenario.