Remove user_catalog_table=true from spock.progress
The new table for tracking progress, used in commit ordering,
does not need the user_catalog_table=true setting, as it is
never going to be used inside the output-plugin. We are not
changing any other tables that may also have this flag set by
copy/paste mistake at this point to error on the safe side.