This patch is a WIP.
It fixes the issue with blocked apply workers. Apply workers sleeping
on the conditional variable latch wakeup every 1ms and check if
any other apply worker is blocked by them. If so, we create the
necessary log entries and exit.

As part of this patch, the suffix is added to apply worker processes
that should help in debugging commit ordering. May be this should
only be done when running PG in debug mode.

The apply works at the moment may all end up waiting for the latch.
It was to be set by the worker that detected the conflict and exited.