Call pg_replication_origin_xact_setup() on sync
We need to make the data sync COPY to treat data as if it
was created at EPOCH, so that incoming transactions that
happened after the slot creation snapshot happened will
always win the LWW conflict resolution. This now makes a
Zero-Downtime-Add-Node work for creating the first replica
for what was a single, stand-alone database before.