Queue stream messages to replay on exception
Create a queue that collects replication stream messages up to
a GUC (spock.exception_replay_queue_size) number of bytes. In
case the original transaction (before exception handling got
activated) ERRORs out, the apply-worker will not terminate, but
internally reset and retry the same transaction in exception
handling mode, replaying the messages from memory without
tearing down the replication connection.

Note: The apply-worker previously created a ResourceOwner. This
seems completely obsolete since it is done in StartTransaction().
It also interfered with resource cleanup while handling the
initial ERROR.

Note: The current queue limit of 4MB it totally arbitrary. Since
this is an allocation maximum on an apply-worker, it doesn't
seem excessive. It is meant to prevent OOM on gigantic
transactions by forcing a restart only if a transaction of that
size causes an exception.