Fix wait_for_sync_event() by making it a stored proc
The upgrade procedure from v4.0.x to current was broken. The
function as implemented caused bloat on the entire database
because it required an MVCC snapshot to be active for the
entire duration.

This commit fixes both issues by turning wait_for_sync_event()
into a stored procedure and dropping the lag_tracker view
before attempting to re-create the underlying function during
the upgrade process.