Enhance synchronous replication feedback handling to avoid premature acknowledgment.
When a synchronous replica is attached, the system previously set synchronous_commit
to off, which could lead to premature feedback being sent before receiving acknowledgment
from the remote synchronous replica. This behavior posed a risk to data consistency.

To address this, a list of locally committed LSNs is maintained. Feedback is now delayed
until acknowledgment is received from the remote synchronous replica, ensuring transaction
durability while preventing unnecessary blocking.

Now the the feedback will be send if 1/2 the wal_sender_timeout occurs or in case of 10 'w'
messages.