[SPOC-90] Fixed triggers regression test (#98)
* Fix replicate_ddl default search_path with updated behavior

This was inconsistent with the overloaded version of the function,
which correctly defaulted to current_setting('search_path'). This
patch updates the original version to match, resolving the issue.

Updated triggers.out expected file to reflect schema-less NOTICE
messages (due to search_path being set).

* Updated regression test expected files to reflect changes in NOTICE
and ERROR output, which no longer include schema-qualified names. This
is due to the updated search_path default in the spock.replicate_ddl
function.