Let the backend remember the state of REPAIR mode.
Previously, the backend had ignored the REPAIR state, simply writing a WAL
message and switching the flag within the walsender process. As a result, in
the repair mode, it operated similarly to the normal mode, adding DDLs to
the queue, sending client messages, etc.

It looks a little misleading. So, by this commit, we switch
the spock_replication_repair_mode on the backend as well as inside
the walsender. It is implemented as an internal session GUC that restores its
state to the default at the end of the transaction or on the next call of
the 'spock_repair_mode' function.
Regression tests have been added.

Because the function may influence replication consistency, a LOG message
was introduced to indicate that the mode status was shifted.

XXX: Does this approach ensure that the backend's and walsender's setting of
the repair_mode flag will behave similarly?

(cherry picked from commit e00bafe7a9c6618b1d9e9f1f70ca02d8390dbd13)