Include the spock_compat.h file locally
It seems there is a little mess with the compatibility macros. It is intended
to avoid multiple #ifdef insertions by adding global macros. But without care,
it may interfere with the core macros and functions. So, it would be safer
to include compatibility headers inside the module that needs them.