Wipe out password from error message.
To prevent password exposure in the log file, examine each ERROR message
emitted by Postgres processes (loaded Spock module) and wipe out symbols from
the error message and query string following the keyword 'password'.

Other c-strings exist in the ErrorData structure. Can it happen if the password
is included in these strings and is written down in the log?
As a negative outcome, we may possibly wipe out a valuable part of
the ERROR message.

This commit introduces a message filter hook that enables Spock to get rid of
annoying messages effectively. This may allow us to remove one of
the core patches.