Follow the exception rule for INSERT in a non-existing table (#102).
A user may manually add a local table to the replication set. In this case,
any DML on this table will be replicated to subscribers. Such a record
shouldn't halt the applying worker, but must follow the standard exception rule.

With this commit, we add processing of the 'Relation does not exist' error.
Since table absence is detected without error, a slight modification was made
to the log_insert_exception function interface: instead of edata, we now pass
only the error message.

Add to the LOG message 'SPOCK: caught exception while use_try_block=true'
description of the actual error that caused the apply worker halting.

NOTE: UPDATE and DELETE also should be handled that way.