SPOC-102 - INSERT into an absent table shouldn't halt the apply worker (#168)
* Follow the exception rule for INSERT in a non-existing table (#102).

A user may manually add a local table to the replication set. In this case,
any DML on this table will be replicated to subscribers. Such a record
shouldn't halt the applying worker, but must follow the standard exception rule.

With this commit, we add processing of the 'Relation does not exist' error.
Since table absence is detected without error, a slight modification was made
to the log_insert_exception function interface: instead of edata, we now pass
only the error message.

Add to the LOG message 'SPOCK: caught exception while use_try_block=true'
description of the actual error that caused the apply worker halting.

Also remove one of the Postgres patch files for all version and instead
make use of the logging hook.

* Follow the exception rule for UPDATE and DELETE in a non-existing table

* Add tests to check how Spock processes the case when the apply worker can't
process an INSERT because of the absent relation (#102).

(cherry picked from commit 7ff1480502d71917057223590a0c46bd31b904cf)