SPOC-51: obfuscate DSN password in error logs (#158)
* Wipe out password from error message.

To prevent password exposure in the log file, examine each ERROR message
emitted by Postgres processes (loaded Spock module) and wipe out symbols from
the error message and query string following the keyword 'password'.

Other c-strings exist in the ErrorData structure. Can it happen if the password
is included in these strings and is written down in the log?
As a negative outcome, we may possibly wipe out a valuable part of
the ERROR message.

This commit introduces a message filter hook that enables Spock to get rid of
annoying messages effectively. This may allow us to remove one of
the core patches.

* Lower log level for a couple of core messages.

Reduce error level from LOG to DEBUG1 for a couple of messages:
1. "tuple to be locked was already moved to another partition due to concurrent update, retrying"
2. "concurrent delete, retrying"

That makes the core patch 'pg15-020-LOG-to-DEBUG1' obsolete.

Also, small fix for the 'password obfuscation' feature to make static code
analysers quiet.