Skip replication of a utility statement containing unsafe DDL.
Some issues may arise when replicating the CREATE/ALTER Spock extension. Also,
some functions may be unsafe to execute on a remote instance. To address this
issue, we propose an approach to identify such statements and instruct Spock
not to replicate them.

Considering that a problematic statement may lie deeply within the statement -
for example, inside the function - we control recursion in the utility
statement tree and raise a flag if something unsafe is detected.

We detect unsafe statements by statement type and their content. It is not
the ideal approach, but the cheapest and simplest one.

NOTE:
Another approach exists: Spock may analyse the changed object inside
the 'spock_object_access', check the object's dependencies
(findDependentObjects), and determine if Spock is in this list. It basically
works and provides very detailed control over objects. However, it is
obviously more expensive than proposed in this commit. It may be implemented
later if necessary.

Also, do some code readability stuff, just in-place.