Bugfix on cleared memory access at the COMMIT callback.
There is the signal_workers list that should be processed at the end of
the transaction. It existed in TopTransactionContext, and there was a case
where the transaction was aborted or prepared. Upon abort, all resources
are released, and signal_workers should be freed and nullified.

By this commit, we add a proper reaction to each XactEvent. Also, remove
the 'xacthook_signal_workers' flag because the 'signal_workers' NULL value
covers this logic.
The signal_workers list is now allocated in the TopMemoryContext. It may hide
some issues, but it simplifies the code in the case of PREPARE TRANSACTION.
Also, some regression tests were added.

NOTE:
It seems that the entire concept of changing the set of subscriptions may be
revised: instead of modifying the Spock catalogue within the backend, we could
perform all the necessary tasks within the worker manager. Just put
the required data into the shared memory and send a signal to the workers
manager. At the same time, we would need to wait until the manager finishes
the operation and sends the 'end' signal back to the backend.