Skip replicating a TRUCATE record during a table resynchronisation.
It is a local operation that resets the table before initiating the syncing.
Use the same machinery as repair_mode and introduce
the skip_wal_records_decoding routine, which can be used everywhere to protect
specific commands. Currently, it is used in repair mode and inside the RESYNC
to truncate the table locally.

New regression test 'sync_table' has been introduced.