Fix two bugs in the table re-sync routine.
At first, we need to switch off WAL sending machinery when truncating and
re-synchronising the table. It prevents us from losing data across
the instances.

The second issue is related to the infinite waiting if no more DML is committed
to the table. The SYNC apply worker report status only on commit. So,
the applying worker, who has initiated this synchronisation, should provide
the SYNC worker not with the last received record LSN, but with
the last committed LSN.

A simple regression test set is provided.

Being in the context, revisit the documentation and visual representation of
the re-sync routine in the UI script.