Increase stability of regression tests.
There are a few tests that rely on the time period required to synchronise
a subscription or deliver replicated data to a subscriber. Such time dependency
causes tests' instability, which can be easily checked by inserting something
like 'pg_usleep(10000);' into any part of the Postgres code where
the confirmed_flush parameter is advanced.

With this commit, we utilise standard Spock machinery to wait for the WAL
flushing event, as well as to perform infinite cycles waiting for a change
in subscription state.

It seems an infinite wait is not the best choice. In the future, we could
introduce a 'wait_change_timeout' environment parameter that can be configured
according to the characteristics of the test platform.

One more race we have detected here is that, altering the subscription, Spock
just sends a signal to the apply worker. So, switching to the publisher quickly
enough, we may find that the replication slot is still not ready to send data.
Just wait until the plugin is ready - yeah, still in an infinite loop.