Generate subscription ID (#192)
* Generate subscription ID

Emulate behaviour of Oid type: don't allow negative and zero values here.
In principle, the sub_id type should be changed to int32 or int64. However,
it would immediately cause numerous changes throughout the code. And what's
worse, it will cause the Spock UI change, which may need changes to the CLI and
other components...
Therefore, simply change the value generator from hash_any to a safer and more
stable sequence.

Add sequence 'sub_id_generator' in the UI of Spock.

The output of some tests was changed because of uncovered sub_id values.
It would be better to hide this value as unstable by definition, but let it go
for now to simplify the commit and keep in mind the good enough stability
of the sequence.

* Make init.sql regression test more stable across versions.

To identify default spock settings don't use \dx command, but directly query
pg_extension and pg_description system tables. So, this test should not change
its output after switching to another Spock version.