Introduce a transient Docker image (SPOC-250). (#216)
* Introduce a transient Docker image.

To avoid repeating the process of refreshing the OS image required for Spock
testing, prepare it once, store it at ghcr.io, and reuse it to build a Docker
image with a patched Postgres and Spock extension.

No automatic check of the cached entry is necessary; re-run the workflow
manually in GitHub Actions if the image appears slightly stale.

TODO:
It makes sense to cache a Docker image with a patched Postgres branch for
supported versions on top of a basic Docker image, as we don't frequently
change our patches. There will be some checks necessary to ensure that
the patch set hasn't been changed since the last image creation, as well as
the 'install.py' script.

(cherry picked from commit 56c362dd4b9c049c06c8f876afb09af9cf2d22b7)