Teach apply_group_progress to return NULL if the timestamp is not set yet.
There was a problem with a new node that showed an oversized lag, because
a zero timestamp is interpreted as an actual time.
Now we treat zero value as 'not set' and return NULL to let the user identify
such an initial state.
Consequently, the spock.progress and spock.lag_tracker views can process NULL
to return a comprehensible result.

Fix the adjust_progress_info bug:
Shifting progress to HTAB, we forgot that now, selecting from the lag_tracker
table, we see records from all databases. So, here we select only the progress
state related to the current database.