Teach apply_group_progress to return NULL if the timestamp is not set yet.
There was a problem with a new node that showed an oversized lag, because
a zero timestamp is interpreted as an actual time.
Now we treat zero value as 'not set' and return NULL to let the user identify
such an initial state.
Consequently, the spock.progress and spock.lag_tracker views can process NULL
to return a comprehensible result.