Update remote_insert_lsn more frequently.
With this commit, the remote_insert_lsn value is upgraded almost as frequently
as received_lsn - each spock plugin WAL record supplied by the report on
the last inserted lsn. It is important, because the WAL may be full of another
transactions that will never be sent to the apply worker. So, apply worker may
get stuck into the situation when never has happened, we see zero lag, but
there are still plenty of transactions to be committed.

NOTE:
We replace highly contended GetXLogInsertRecPtr with much more relaxed
GetXLogWriteRecPtr() - since ee1cbe8 it reads an atomic value
avoiding to acquire spinlock.

To reduce overhead even more we may request detailed statistics at the moment
of replication slot creation - by an output plugin parameter. Another option -
introduce xact_callback and register each COMMIT LSN or TS in Spock's
independent atomic values.

It lets users and automation tools rely on this value to analyse replication
lag or decide on data consistency. The profit we aim to achieve immediately:
users of the local instance don't need to visit the remote node to capture
the replication lag.