Fixes after review.
Remove check_and_update_progress because it doesn't do anything useful now and
anyway, it was too brave changing data in shared memory without locking.

Replace comment for prev_remote_ts: 'obsolete' term is not fully correct because
formally it is used in the code. Just reconsider it in the future.