Let the 'spock.progress' VIEW show data for the current database only.
After switching the resource manager to HTAB, the apply_group_progress routine
exposes the progress state for all the databases. It may be helpful to itself,
but the spock.progress, being database-local, has no reason to show all
the data. Hence, restrict it to return only the current database rows.

As a bonus, it makes Spock's internal code a little simpler and safer.
To prevent future issues, add checks to adjust_progress_info to ensure we view
data for a single database only.