Improve the wait_slot_confirm_lsn routine description.
This function has a special corner case - it waits in an infinite loop until
at least one proper replication slot is found. In Spock, a subscription may be
accidentally disabled due to a critical error that may cause a stall.

Describe the case in more detail and add a test to give users and developers
a clue how to use this routine safely.