zodan.sql: add subscriptions pre-check.
With this commit, we require that each node of the Spock cluster has only
enabled subscriptions. There is no algorithmic reason to do that, because when
adding the node, we are concerned only about the donor's state and
the consistency of its progress table. However, such a strict procedure ensures
that users can be sure they add a node to a healthy cluster, take action
beforehand if it is not healthy, and that replication doesn't delay WAL cutting.
Introduce a TAP test that could be used to check Z0DAN checks.

XXX: add_node finishes its job with subscriptions in the 'initialising' state.
Do we need to take any action here?

Also, fix the annoying WARNING on the 'exception_replay_queue_size'.